Fixes a first-run redirect error
Basically I get a `Cannot read property 'path' of undefined` when the page first loads when it starts by redirecting. More importantly, I must be doing something wrong because `transition` (which is a route instance) never seems to have a `to` property; I scoured the current version of vue-router and couldn't find anything like that (perhaps this is a vestige of Vue 1's router, which I never used?).